pyproject.toml
[tool.poetry]
name = "namixcode-backend"
version = "0.1.0"
description = ""
authors = [
    "lysmux <kllraz@yandex.ru>",
    "XP-T"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standart"], version = "^0.115.0" }
uvicorn = "^0.32.0"
pydantic-settings = "^2.5.2"
dishka = "^1.3.0"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
pydantic = {extras = ["email"], version = "^2.9.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

